[도전과제] 도전과제 완료 - 패키지 리팩토링 - 주석 추가 - 어색한 함수 이름 Rename
