+++ app/src/main/java/org/sopt/sample/HomeActivity.kt
@@ -2,7 +2,7 @@ package org.sopt.sample
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import org.sopt.sample.data.UserData
+import org.sopt.sample.model.UserData
 import org.sopt.sample.databinding.ActivityHomeBinding
 
 class HomeActivity : AppCompatActivity() {
@@ -16,13 +16,13 @@ class HomeActivity : AppCompatActivity() {
         getResult()
     }
 
-    private fun getResult() {
-        val userData =
+    private fun getResult() { // LoginActivity가 보낸 intent 수신
+        val userData = //getParcelableExtra 함수 deprecated 대처
             if (android.os.Build.VERSION.SDK_INT >= 33)
                 intent.getParcelableExtra("userdata", UserData::class.java)!!
             else
                 intent.getParcelableExtra("userdata")!!
-        binding.userdata = userData
+        binding.userdata = userData // 데이터바인딩
     }
 
 }
