+++ app/src/main/java/org/sopt/sample/SignUpActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
-import org.sopt.sample.data.UserData
+import org.sopt.sample.model.UserData
 import org.sopt.sample.databinding.ActivitySignUpBinding
 
 class SignUpActivity : AppCompatActivity() {
@@ -22,9 +22,9 @@ class SignUpActivity : AppCompatActivity() {
         val intent = Intent(this@SignUpActivity, LoginActivity::class.java)
         binding.apply {
             btnSignUp.setOnClickListener {
-                if (etId.text.length in 6..10 && etPw.text.length in 8..12) {
+                if (etId.text.length in 6..10 && etPw.text.length in 8..12) { //회원가입 조건 ID 6~10 / 비밀번호 8~12
                     intent.apply {
-                        putExtra(
+                        putExtra( //UserData 형태로 Intent 전달
                             "userdata",
                             UserData(
                                 etId.text.toString(),
@@ -37,6 +37,7 @@ class SignUpActivity : AppCompatActivity() {
                     finish()
                 } else {
                     Snackbar.make(binding.root, "회원가입이 불가 합니다..", Snackbar.LENGTH_SHORT).apply {
+                        anchorView = binding.btnSignUp
                     }.show()
 
                 }
