+++ app/src/main/java/org/sopt/sample/model/UserData.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.data
+package org.sopt.sample.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
