+++ app/src/main/res/layout/activity_home.xml
@@ -1,15 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout>
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
 
     <data>
 
         <variable
             name="userdata"
-            type="org.sopt.sample.data.UserData" />
+            type="org.sopt.sample.model.UserData" />
     </data>
 
-    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
+    <androidx.constraintlayout.widget.ConstraintLayout xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -21,7 +20,8 @@
             android:layout_height="0dp"
             android:layout_marginHorizontal="150dp"
             android:layout_marginVertical="100dp"
-            android:src="@drawable/drawable_introuce"
+            android:contentDescription="@string/introduce"
+            android:src="@drawable/drawable_junwookim"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -29,25 +29,25 @@
 
         <TextView
             android:id="@+id/tv_name"
+            style="@style/Home_TextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginVertical="12dp"
             android:text="@{`이름 : ` + userdata.id}"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/imageView"
-            tools:text="이름:김준우" />
+            tools:text="이름 : 김준우" />
 
         <TextView
             android:id="@+id/tv_mbti"
+            style="@style/Home_TextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginVertical="12dp"
             android:text="@{`MBTI : ` + userdata.mbti}"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/tv_name"
-            tools:text="MBTI:INFP" />
+            tools:text="MBTI : INFP" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
